+++ lib/dynamodb/actions/save.js
@@ -1,6 +1,6 @@
 import { normalizeProject } from '../tables/projects';
 
-export const save = (dynamo, table, data, cb = () => {}) => {
+export const save = ({ dynamo, table, data }, cb = () => {}) => {
   const params = {
     TableName: table,
     Item: data,
@@ -15,7 +15,25 @@ export const save = (dynamo, table, data, cb = () => {}) => {
   });
 };
 
-export const saveProject = (dynamo, table, project, cb = () => {}) =>
-  save(dynamo, table, normalizeProject(project), cb);
+export const saveProject = ({ dynamo, table, event, data }, cb = () => {}) =>
+  save(
+    {
+      dynamo,
+      table,
+      data: normalizeProject(
+        Object.assign(
+          {
+            creation_date: event.eventTime, // already ISO-8601
+            source: {
+              producer: event.userIdentity.principalId,
+              object_key: event.object.key,
+            },
+          },
+          data
+        )
+      ),
+    },
+    cb
+  );
 
 export default save;
