+++ lib/dynamodb/helpers/normalize.js
@@ -1,11 +1,16 @@
-export const isStringNotEmpty = source =>
-  Object.prototype.toString.call(source) === '[object String]' &&
-  source.trim().length > 0;
+/**
+ * Test if it is a string and if the string is not empty.
+ * @param  {string}  str The string to test.
+ * @return {boolean}     Return true if it is a string and if the string is not empty.
+ */
+export const isStringNotEmpty = str =>
+  Object.prototype.toString.call(str) === '[object String]' &&
+  str.trim().length > 0;
 
 /**
  * Normalize string for DynamoDB.
  * @param  {string} source The string to normalize.
- * @return {object}        The normalized string object or null if the source is not a string or if it is empty.
+ * @return {Object}        The normalized string object or null if the source is not a string or if it is empty.
  */
 export const normalizeString = source =>
   isStringNotEmpty(source)
@@ -17,7 +22,7 @@ export const normalizeString = source =>
 /**
  * Normalize number for DynamoDB.
  * @param  {number} source The number to normalize.
- * @return {object}        The normalized number object or null if the source is not a finite number.
+ * @return {Object}        The normalized number object or null if the source is not a finite number.
  */
 export const normalizeNumber = source =>
   Number.isFinite(source)
@@ -28,8 +33,8 @@ export const normalizeNumber = source =>
 
 /**
  * Normalize string set for DynamoDB.
- * @param  {array} source The string set to normalize.
- * @return {object}       The normalized string set or null if the source is not a string set or if it is empty.
+ * @param  {string[]} source The string set to normalize.
+ * @return {Object}          The normalized string set or null if the source is not a string set or if it is empty.
  */
 export const normalizeStringSet = source => {
   if (Array.isArray(source) && source.length > 0) {
@@ -47,10 +52,20 @@ export const normalizeStringSet = source => {
   return null;
 };
 
+/**
+ * Normalize a map for DynamoDB.
+ * @param  {Object} source The object to normalize.
+ * @return {Object}        The normalized object.
+ */
 export const normalizeMap = source => ({
   M: source,
 });
 
+/**
+ * Normalize a list for DynamoDB.
+ * @param  {Array}  source The array to normalize.
+ * @return {Object}        The normalized array or null if the source is not an array or if it is an empty array.
+ */
 export const normalizeList = source =>
   Array.isArray(source) && source.length > 0
     ? {
